fix(root): change targets for skip links and add tabindex="-1" to get virtual cursor working
For "Skip to main content" link, add <a> with tabindex of -1 as target, and for "Skip to page
content" link, replace page header heading with slot and set that as the target with tabindex of -1.
Having a tabindex of -1 helps with focus - particularly getting the virtual cursor to move. Add
eslint disable comment for <a>.

.#616