fix(root): fix errors picked up by A11y Insights FastPass and WAVE tool
Fix errors which were appearing when you run Accessibility Insights' FastPass mode and also the WAVE
plug-in. These errors were due to the icons / logos not having discernible text because of the
"aria-labelledby" not being used quite correctly. Set aria-hidden on org logos as the link text was
already being read out due to the span text. Add full stops to "SIS" so it's a bit clearer for
screen reader users.